requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.5
future==0.17.1
plac==1.0.0
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
termcolor==1.1.0
tinydb==3.12.2
